% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{approx_pwr}
\alias{approx_pwr}
\title{approx_pwr: approximate power calculations}
\usage{
approx_pwr(b0, b1, b2, rho, n, alpha = 0.05, dist = "normal", mu = 0,
  sigma = 1)
}
\arguments{
\item{b0}{hypothesized intercept of the logistic regression model}

\item{b1}{hypothesized effect size of variable of interest X1}

\item{b2}{hypothesized effect size of adjustment variable X2}

\item{rho}{hypothesized correlations between variable of interest X1 and adjustment variable X2}

\item{n}{current sample size}

\item{alpha}{significance level}

\item{dist}{supposed data distribution. only supports "normal" at the moment.}

\item{mu}{mean of the supposed normal distribution of X1}

\item{sigma}{standard deviation of the supposed normal distribution of X1}
}
\value{
approximate power
}
\description{
approx_pwr: approximate power calculations
}
