% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{approx_samp_size}
\alias{approx_samp_size}
\title{approx_samp_size: approximate sample size given multivariate-normal distribution and its parameters}
\usage{
approx_samp_size(b0, b1, b2, rho, targetpwr, alpha, dist = "normal",
  mu = 0, sigma = 1)
}
\arguments{
\item{b0}{hypothesized intercept of the logistic regression model}

\item{b1}{hypothesized effect size of variable of interest X1}

\item{b2}{hypothesized effect size of adjustment variable X2}

\item{rho}{hypothesized correlations between variable of interest X1 and adjustment variable X2}

\item{targetpwr}{power target, by default = 0.80}

\item{alpha}{significance level}

\item{dist}{supposed data distribution. only supports "normal" at the moment.}

\item{mu}{mean of the supposed normal distribution of X1}

\item{sigma}{standard deviation of the supposed normal distribution of X1}
}
\value{
approximate sample size
}
\description{
approx_samp_size: approximate sample size given multivariate-normal distribution and its parameters
}
