% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{calc_pwr_emp}
\alias{calc_pwr_emp}
\title{calc_pwr_emp: approximate power from design matrix}
\usage{
calc_pwr_emp(b0, b1, b2 = NULL, xmat, alpha = 0.05, reg = "uni")
}
\arguments{
\item{b0}{hypothesized intercept of the logistic regression model}

\item{b1}{hypothesized effect size of variable of interest X1}

\item{b2}{hypothesized effect size of adjustment variable X2}

\item{xmat}{design matrix}

\item{alpha}{significance level}

\item{reg}{either "uni" for univaraite calculations or "multi" for bivariate calculations}
}
\value{
approximate power from design matrix
}
\description{
calc_pwr_emp: approximate power from design matrix
}
