% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{calc_samp_size}
\alias{calc_samp_size}
\title{calc_samp_size: approximate sample size given data distribution and its parameters}
\usage{
calc_samp_size(b0, b1, targetpwr, alpha, dist = "normal", mu = 0,
  sigma = 1, lwr = 0, upr = 1, ctr = 0, scl = 1, df = 10,
  succ = 1, fail = 0, p = 0.5)
}
\arguments{
\item{b0}{hypothesized intercept of the logistic regression model}

\item{b1}{hypothesized effect size of variable of interest X1}

\item{targetpwr}{power target, by default = 0.80}

\item{alpha}{significance level}

\item{dist}{supposed data distribution}

\item{mu}{the mean of supposed normal distribution}

\item{sigma}{the standard deviation of supposed normal distribution}

\item{lwr}{the lower bound of supposed uniform distribution}

\item{upr}{the upper bound of supposed uniform distribution}

\item{ctr}{the center of supposed t distribution}

\item{scl}{the scale of supposed t distribution}

\item{df}{the degree of freedom of supposed t distribution}

\item{succ}{the value for "success" in a supposed binary distribution, usually 1}

\item{fail}{the value for "failure" in a supposed binary distribution, usually 0}

\item{p}{the proportion of "successes" in a supposed binary distribution}
}
\value{
approximate sample size
}
\description{
calc_samp_size: approximate sample size given data distribution and its parameters
}
