% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{calc_samp_size_emp}
\alias{calc_samp_size_emp}
\title{calc_samp_size_emp: approximate sample size from design matrix}
\usage{
calc_samp_size_emp(xmat, b0, b1, b2 = NULL, targetpwr = 0.8,
  alpha = 0.05, reg = "uni")
}
\arguments{
\item{xmat}{design matrix}

\item{b0}{hypothesized intercept of the logistic regression model}

\item{b1}{hypothesized effect size of variable of interest X1}

\item{b2}{hypothesized effect size of adjustment variable X2}

\item{targetpwr}{power target, by default = 0.80}

\item{alpha}{significance level}

\item{reg}{either "uni" for univaraite calculations or "multi" for bivariate calculations}
}
\value{
approximate sample size from design matrix
}
\description{
calc_samp_size_emp: approximate sample size from design matrix
}
