% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{findb0}
\alias{findb0}
\title{findb0: find the intercept in a logistic model}
\usage{
findb0(b1, prop, dist = "normal", b2 = NULL, multi = F, mu = 0,
  sigma = 1, lwr = 0, upr = 1, ctr = 0, scl = 1, df = 10,
  succ = 1, fail = 0, p = 0.5)
}
\arguments{
\item{b1}{hypothesized effect size}

\item{prop}{A number between 0 and 1 that represents case/success/outcome==1 proportion}

\item{dist}{supposed data distribution}

\item{b2}{hypothesized adjustment variable effect size, if multi}

\item{multi}{univariate if == F, multivariate if == T}

\item{mu}{the mean of supposed normal distribution}

\item{sigma}{the standard deviation of supposed normal distribution}

\item{lwr}{the lower bound of supposed uniform distribution}

\item{upr}{the upper bound of supposed uniform distribution}

\item{ctr}{the center of supposed t distribution}

\item{scl}{the scale of supposed t distribution}

\item{df}{the degree of freedom of supposed t distribution}

\item{succ}{the value for "success" in a supposed binary distribution, usually 1}

\item{fail}{the value for "failure" in a supposed binary distribution, usually 0}

\item{p}{the proportion of "successes" in a supposed binary distribution}
}
\value{
calculated intercept value in the logistic model given
the effect size \code{b1}, proportion \code{prop} and distribution parameters \code{mu}, \code{sigma} or so forth.
}
\description{
findb0: find the intercept in a logistic model
}
\examples{
findb0(b1 = 0.33, prop = 0.5, dist = "t", ctr = 3, scl = 6, df = 9)
}
