% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{findb0emp}
\alias{findb0emp}
\title{findb0emp: find the intercept in a logistic model from an emperical distribution}
\usage{
findb0emp(b1, b2 = NULL, prop, xmat, reg = "uni", lower, upper)
}
\arguments{
\item{b1}{hypothesized effect size}

\item{b2}{hypothesized adjustment variable effect size}

\item{prop}{A number between 0 and 1 that represents case/success/outcome==1 proportion}

\item{xmat}{design matrix}

\item{reg}{either "uni" for univaraite calculations or "multi" for bivariate calculations}

\item{lower}{lower bound of the hypothesized effect size b1}

\item{upper}{upper bound of the hypothesized effect size b1}
}
\value{
calculated intercept value in the logistic model given
the hypothesized effect sizes \code{b1} and possibly \code{b2},
proportion \code{prop} and design matrix \code{xmat},
with \code{reg} specifying the a univariate or multivariate regression model,
seaching from \code{lower} to \code{upper} of the hypothesized effect size \code{b1}.
}
\description{
findb0emp: find the intercept in a logistic model from an emperical distribution
}
\examples{
findb0emp(b1 = 0.33, b2 = 0.13, prop = 0.5, xmat = design_matrix, reg = "multi", lower = 0.03, upper = 0.83)
}
