% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{get_pwr}
\alias{get_pwr}
\title{get_pwr: approximate power calculations given information matrix}
\usage{
get_pwr(i0, i1, i2, b1, b2 = NULL, rho = NULL, n, alpha, reg)
}
\arguments{
\item{i0}{I11 in the information matrix of esimated b1}

\item{i1}{I12 and I21 in the information matrix}

\item{i2}{I22 in the information matrix}

\item{b1}{hypothesized effect size}

\item{b2}{hypothesized adjustment variable effect size, if reg == "multi"}

\item{rho}{hypothesized correlations between variable of interest X1 and adjustment variable X2}

\item{n}{current sample size}

\item{alpha}{significance level}

\item{reg}{either "uni" for univaraite calculations or "multi" for bivariate calculations}
}
\value{
calculated power given information matrix
}
\description{
get_pwr: approximate power calculations given information matrix
}
