% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlog.R
\name{get_samp_size}
\alias{get_samp_size}
\title{get_samp_size: approximate sample size}
\usage{
get_samp_size(i0, i1, i2, b1, b2 = NULL, rho = NULL, targetpwr = 0.8,
  alpha = 0.05, reg = "uni")
}
\arguments{
\item{i0}{I11 in the information matrix of esimated b1}

\item{i1}{I12 and I21 in the information matrix}

\item{i2}{I22 in the information matrix}

\item{b1}{hypothesized effect size between 0 to 10000}

\item{b2}{hypothesized adjustment variable effect size, if reg == "multi"}

\item{rho}{hypothesized correlations between variable of interest and adjument variable.}

\item{targetpwr}{target power, default to 0.80}

\item{alpha}{significance level}

\item{reg}{either "uni" for univaraite calculations or "multi" for bivariate calculations}
}
\value{
calculated power given parameters
}
\description{
get_samp_size: approximate sample size
}
